Fix build for cfg(async_payments)
Static invoices don't have an amount_msats field.
